extract VariantType

+++ src/main/java/uk/ac/ebi/eva/commons/models/mongo/entity/Annotation.java
@@ -19,6 +19,7 @@ import com.google.common.base.Strings;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
+import uk.ac.ebi.eva.commons.models.data.Variant;
 import uk.ac.ebi.eva.commons.models.mongo.entity.subdocuments.ConsequenceType;
 import uk.ac.ebi.eva.commons.models.mongo.entity.subdocuments.Xref;
 
@@ -53,10 +54,10 @@ public class Annotation {
     private String chromosome;
 
     @Field(value = START_FIELD)
-    private Long start;
+    private int start;
 
     @Field(value = END_FIELD)
-    private Long end;
+    private int end;
 
     @Field(value = VEP_VERSION_FIELD)
     private String vepVersion;
@@ -74,7 +75,7 @@ public class Annotation {
         // Empty document constructor for spring-data
     }
 
-    public Annotation(String chromosome, Long start, Long end, String referenceAllele, String alternativeAllele,
+    public Annotation(String chromosome, int start, int end, String referenceAllele, String alternativeAllele,
                       String vepVersion, String vepCacheVersion) {
         this.chromosome = chromosome;
         this.start = start;
@@ -111,11 +112,11 @@ public class Annotation {
         return chromosome;
     }
 
-    public Long getStart() {
+    public int getStart() {
         return start;
     }
 
-    public Long getEnd() {
+    public int getEnd() {
         return end;
     }
 
@@ -162,7 +163,7 @@ public class Annotation {
         }
     }
 
-    public static String buildAnnotationId(String chromosome, Long start, String reference, String alternate,
+    public static String buildAnnotationId(String chromosome, int start, String reference, String alternate,
                                            String vepVersion, String vepCacheVersion) {
         StringBuilder builder = new StringBuilder(VariantDocument.buildVariantId(
                 chromosome, start,
